+++ app/(memorybook)/memorybook/(itinerary)/[trip_id]/layout.tsx
@@ -1,10 +1,10 @@
 import { Metadata } from "next";
+import { notFound } from "next/navigation";
 
+import { getTrip } from "../../lib/memoryBookService";
 import FooterMenu from "../../components/layout/footer/FooterMenu";
 import Header from "../../components/layout/header/Header";
 import Footer from "../../components/layout/footer/Footer";
-import { getTrip } from "../../lib/memoryBookService";
-import { notFound } from "next/navigation";
 import MainContainer from "@/app/components/layout/MainContainer";
 
 export const metadata: Metadata = {
@@ -31,7 +31,7 @@ export default async function RootLayout({
     <>
       <Header />
       <main className="flex-1 px-2">
-        <MainContainer>
+        <MainContainer itineraryPage={true}>
           <div className="px-1 md:px-8 w-full">
             <h2 className="mt-0 md:mt-8 p-2 text-lg md:text-2xl text-center border-b border-solid text-black border-blue-800 bg-white">
               {trip?.name}
