+++ app/(memorybook)/memorybook/components/layout/header/Header.tsx
@@ -19,7 +19,7 @@ const Header = async () => {
     <header className="w-full">
       <div className="max-w-[1120px] mx-auto">
         <div className="flex justify-between items-center mx-2 h-16">
-          <Link href="/memorybook">
+          <Link href="/">
             <h1>
               <Image
                 src="/logo_itinerary.png"
